+++ extensions/blocks/conversation/index.js
@@ -11,7 +11,7 @@ import { ConversationIcon as icon } from '../../shared/icons';
 /**
  * Local dependencies
  */
-import './extensions';
+import './extend';
 import attributes from './attributes';
 import edit from './edit';
 import save from './save';
